update node, port and UI input

+++ Assets/_Project/Scripts/VisualScripting/OperatorHandle/AdditionHandler.cs
@@ -3,8 +3,8 @@ public class AdditionHandler : OperatorHandler
     public AdditionHandler()
     {
         Operator<string, string>((a, b) => a + b);
-        Operator<double, double>((a, b) => a + b);
-        Operator<string, double>((a, b) => a + b);
-        Operator<double, string>((a, b) => a + b);
+        Operator<float, float>((a, b) => a + b);
+        Operator<string, float>((a, b) => a + b);
+        Operator<float, string>((a, b) => a + b);
     }
 }
